fix : file moved to data dir
